Fix TCP connection not persisting
This issue was causing the client to connect and immediately disconnect
because the server was not looping on the socket data, closing the
connection after just one message

Fixed by adding a loop to the socket read. This caused a second issue
that the login connection is kept open by the client, blocking the
server. So I implemented asynchronous TCP listening based on [1].

[1]: https://rust-lang.github.io/async-book/09_example/01_running_async_code.html